Add a script to retrieve all systems running with domains
This is a convenience script for operators to run to retrieve all systems running in the cloud.gov platform that have a domain associated with them.  The purpose of this script is to make it easier to pull system metrics too get an estimate of what is currently running.

The script assumes you are logged into CF in the command line, and outputs information in CSV format to be put in a spreadsheet.