Improve active sandbox and space count script
This changeset adds a few improvements to the active sandbox and space count script by allowing an operator to specify start and end dates to filter by.  If a start date (and an end date) are given, the results are constrained to those dates in an inclusive fashion: starting at 00:00:00 on the start date and ending at 23:59:59 on the end date (if provided).

Including the dates will also provide a new line of output, how many sandbox spaces were created during that time.